Typechecking tests pass, fixing repl tests
