+++ compiler/src/Language/Mimsa/Logging.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE BangPatterns #-}
+
 module Language.Mimsa.Logging where
 
 import qualified Data.Text as T
@@ -9,4 +11,5 @@ debugLog title item = snd (traceShowId (title, item))
 
 debugPretty :: (Printer b) => String -> b -> b
 debugPretty title item =
-  trace (title <> ":\n" <> T.unpack (prettyPrint item)) item
+  let !output = title <> ":\n" <> T.unpack (prettyPrint item)
+   in trace output item
