+++ compiler/test/golden/PrettyPrint/16af8b7998e76b854967b27bbaf4b9691770d6be6547039265e03c26805442e6.mimsa
@@ -1,0 +1,1 @@
+type Justthing    = String in True
