+++ compiler/test/golden/PrettyPrint/25ae053535573258de073f50e792f1e09803ee195b4023807e4977735afbad17.mimsa
@@ -1,0 +1,8 @@
+let fromMaybe =
+  \def ->
+    \maybe ->
+      match maybe with 
+          (Just a) -> a
+        | (Nothing) -> def;
+
+fromMaybe("Horse")((Just 1))
