+++ compiler/test/golden/StoreExpr/06ec8e191bde022d503e9b68d4f3abf9f414f219bcc14c09c1ae828e8a7da14d.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"tag":"MyLet","contents":[[],"a",{"tag":"MyRecord","contents":[[],{"one":{"tag":"MyLiteral","contents":[[],{"tag":"MyInt","contents":1}]}}]},{"tag":"MyLet","contents":[[],"one",{"tag":"MyRecordAccess","contents":[[],{"tag":"MyVar","contents":[[],"a"]},"one"]},{"tag":"MyLet","contents":[[],"two",{"tag":"MyRecordAccess","contents":[[],{"tag":"MyVar","contents":[[],"a"]},"two"]},{"tag":"MyVar","contents":[[],"a"]}]}]}]},"storeBindings":{},"storeTypeBindings":{}}
