+++ compiler/test/golden/StoreExpr/16af8b7998e76b854967b27bbaf4b9691770d6be6547039265e03c26805442e6.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"tag":"MyData","contents":[[],{"dtConstructors":{"String":[]},"dtName":"Justthing","dtVars":[]},{"tag":"MyLiteral","contents":[[],{"tag":"MyBool","contents":true}]}]},"storeBindings":{},"storeTypeBindings":{}}
