+++ compiler/test/golden/StoreExpr/25ae053535573258de073f50e792f1e09803ee195b4023807e4977735afbad17.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"tag":"MyLet","contents":[[],"fromMaybe",{"tag":"MyLambda","contents":[[],"def",{"tag":"MyLambda","contents":[[],"maybe",{"tag":"MyPatternMatch","contents":[[],{"tag":"MyVar","contents":[[],"maybe"]},[[{"tag":"PConstructor","contents":[[],"Just",[{"tag":"PVar","contents":[[],"a"]}]]},{"tag":"MyVar","contents":[[],"a"]}],[{"tag":"PConstructor","contents":[[],"Nothing",[]]},{"tag":"MyVar","contents":[[],"def"]}]]]}]}]},{"tag":"MyApp","contents":[[],{"tag":"MyApp","contents":[[],{"tag":"MyVar","contents":[[],"fromMaybe"]},{"tag":"MyLiteral","contents":[[],{"tag":"MyString","contents":"Horse"}]}]},{"tag":"MyConsApp","contents":[[],{"tag":"MyConstructor","contents":[[],"Just"]},{"tag":"MyLiteral","contents":[[],{"tag":"MyInt","contents":1}]}]}]}]},"storeBindings":{},"storeTypeBindings":{"Just":"1e40e1a1c0ddc82e0ddeb7fa0f458b3b954d32f5ae5ccdedffb59de50c7bebac","Nothing":"1e40e1a1c0ddc82e0ddeb7fa0f458b3b954d32f5ae5ccdedffb59de50c7bebac"}}
